+++ wp-content/plugins/astra-sites/inc/lib/ai-builder/inc/compatibility/cartflows/ai-builder-compatibility-cartflows.php
@@ -10,9 +10,7 @@ namespace AiBuilder\Inc\Compatibility\Cartflows;
 
 defined( 'ABSPATH' ) || exit;
 
-
-
-if ( ! class_exists( 'Ai_Builder_Compatibility_Cartflows' ) ) :
+if ( ! class_exists( 'Ai_Builder_Compatibility_Cartflows' ) ) {
 
 	/**
 	 * Cartflows Compatibility
@@ -20,7 +18,6 @@ if ( ! class_exists( 'Ai_Builder_Compatibility_Cartflows' ) ) :
 	 * @since 3.4.6
 	 */
 	class Ai_Builder_Compatibility_Cartflows {
-
 		/**
 		 * Instance
 		 *
@@ -28,7 +25,16 @@ if ( ! class_exists( 'Ai_Builder_Compatibility_Cartflows' ) ) :
 		 * @var object Class object.
 		 * @since 3.4.6
 		 */
-		private static $instance;
+		private static $instance = null;
+
+		/**
+		 * Constructor
+		 *
+		 * @since 3.4.6
+		 */
+		public function __construct() {
+			add_action( 'astra_sites_after_plugin_activation', array( $this, 'disable_cartflows_redirect' ), 2 );
+		}
 
 		/**
 		 * Initiator
@@ -37,29 +43,24 @@ if ( ! class_exists( 'Ai_Builder_Compatibility_Cartflows' ) ) :
 		 * @return object initialized object of class.
 		 */
 		public static function get_instance() {
-			if ( ! isset( self::$instance ) ) {
+			if ( null === self::$instance ) {
 				self::$instance = new self();
 			}
 			return self::$instance;
 		}
 
 		/**
-		 * Constructor
-		 *
-		 * @since 3.4.6
-		 */
-		public function __construct() {
-			add_action( 'astra_sites_after_plugin_activation', array( $this, 'disable_cartflows_redirect' ) );
-		}
-
-		/**
 		 * Disable Cartflows redirect.
 		 *
-		 * @return void.
+		 * @param string               $plugin_init Plugin init file used for activation.
+		 * @param array<string, mixed> $data data.
+		 * @return void
 		 */
-		public function disable_cartflows_redirect() {
-			update_option( 'wcf_setup_skipped', true );
-			delete_option( 'wcf_start_onboarding' );
+		public function disable_cartflows_redirect( $plugin_init, $data = array() ) {
+			if ( 'cartflows/cartflows.php' === $plugin_init ) {
+				update_option( 'wcf_setup_skipped', true );
+				delete_option( 'wcf_start_onboarding' );
+			}
 		}
 	}
 
@@ -68,4 +69,4 @@ if ( ! class_exists( 'Ai_Builder_Compatibility_Cartflows' ) ) :
 	 */
 	Ai_Builder_Compatibility_Cartflows::get_instance();
 
-endif;
+}
